src/auth-service/requirements.txt
Flask
Flask-SQLAlchemy
marshmallow
psutil
kubernetes
boto3
redis
pika 
mysqlclient
flask-cors
flask-jwt-extended
PyJWT